backend/requirement.txt
asgiref==3.5.2
bcrypt==4.0.0
charset-normalizer==2.1.1
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
idna==3.4
mysqlclient==2.1.1
PyJWT==2.4.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.12
